Automated-data-analysis-app/backend/requirements.txt

fastapi
uvicorn
pandas
openpyxl
mysql-connector-python
python-multipart
scikit-learn
matplotlib
seaborn
prophet
weasyprint
xlsxwriter
reportlab
